Fixed #176. The omitted parameter for IO::Normalize was an unsigned int, and the default value was -1. I replaced it with (unsigned int)-1.
git-svn-id: http://svn.ta3d.org/yuni/trunk@988 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/io.h
@@ -257,7 +257,8 @@ namespace IO
 	** \param      inLength Length of #in (optional, -1 for autodetection)
 	*/
 	template<class StringT, class AnyStringT>
-	void Normalize(StringT& out, const AnyStringT& in, unsigned int inLength = -1);
+	void Normalize(StringT& out, const AnyStringT& in,
+		unsigned int inLength = (unsigned int)-1);
